Testing f64 performance again.

+++ src/types.rs
@@ -1,6 +1,6 @@
 
 /// Main type we use in here
-pub type Feature = f32;
+pub type Feature = f64;
 pub type Label = u32;
